change tiem on subject to get from date returned from api

+++ Weather/Data.cs
@@ -2,14 +2,14 @@
 {
     public class Data
     {
-        public int Temperature { get; set; }
+        public double Temperature { get; set; }
         public string WindDirection { get; set; }
         public float WindVelocity { get; set; }
         public double Humidity { get; set; }
         public string Condition { get; set; }
         public float Pressure { get; set; }
         public string Icon { get; set; }
-        public int Sensation { get; set; }
+        public double Sensation { get; set; }
         public string Date { get; set; }
     }
 }
